Refactor API service structure: reorganize service files and update import paths
